requirements.txt
future
emoji
requests
telethon
sqlalchemy
python-telegram-bot==12.7
psycopg2-binary
pythonping
feedparser
pytz
gtts
bs4
lxml
demjson
wget
bs4
telegraph
pillow
psutil
speedtest-cli>=2.0.2
googletrans
wikipedia
spamwatch
bleach
markdown2
alphabet_detector
